Playing around with an idea.
