+++ src/Alert/Alert.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
+
 import { DismissButton } from './DismissButton'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
@@ -9,76 +10,13 @@ import {
   faExclamationTriangle,
   faTimesCircle
 } from '@fortawesome/free-solid-svg-icons'
+
+import { AlertError } from './AlertError'
+import { AlertInfo } from './AlertInfo'
+import { AlertSuccess } from './AlertSuccess'
+import { AlertWarning } from './AlertWarning'
+
 import styles from './alert.module.css'
-
-// Default look and feel
-// /////////////////////////////////////////////////////////////////////////////
-const defaultClasses = {
-  wrapper: 'su-alert',
-  dismissButtonWrapper: [
-    'su-order-3',
-    'su-rs-m-l-1',
-    'su-h-full',
-    'su-items-end',
-    'su-flex-shrink',
-    'su-text-right',
-    'su-w-full',
-    'sm:su-w-auto'
-  ].join(' '),
-  dismissButton: '',
-  headerWrapper: [
-    'su-order-1',
-    'su-rs-m-r-1',
-    'su-flex-shrink',
-    'su-mb-4',
-    'xs:su-w-full',
-    'lg:su-w-max'
-  ].join(' '),
-  label: [
-    'su-inline-block',
-    'su-uppercase',
-    'su-font-semibold',
-    'su-text-170rem',
-    'su-h-full',
-    styles.label
-  ].join(' '),
-  bodyWrapper: styles.alertBodyWrapperDark,
-  icon: faBell,
-  iconClass: 'su-mr-2 su-inline-block su-max-w-xs',
-  bodyHeading: '',
-  footerWrapper: 'su-rs-m-t-0'
-}
-
-// Variant styles.
-// /////////////////////////////////////////////////////////////////////////////
-const variants = {
-  success: {
-    wrapper: 'su-alert su-bg-palo-verde',
-    bodyWrapper: styles.alertBodyWrapper,
-    headerWrapper: defaultClasses.headerWrapper + ' su-text-white',
-    footerWrapper: defaultClasses.footerWrapper + ' su-text-white',
-    icon: faCheckCircle
-  },
-  warning: {
-    wrapper: 'su-alert su-bg-illuminating-dark',
-    bodyWrapper: styles.alertBodyWrapperDark,
-    icon: faExclamationTriangle
-  },
-  error: {
-    wrapper: 'su-alert su-bg-digital-red',
-    bodyWrapper: styles.alertBodyWrapper,
-    headerWrapper: defaultClasses.headerWrapper + ' su-text-white',
-    footerWrapper: defaultClasses.footerWrapper + ' su-text-white',
-    icon: faTimesCircle
-  },
-  info: {
-    wrapper: 'su-alert su-bg-bright-blue',
-    bodyWrapper: styles.alertBodyWrapper,
-    headerWrapper: defaultClasses.headerWrapper + ' su-text-white',
-    footerWrapper: defaultClasses.footerWrapper + ' su-text-white',
-    icon: faQuestionCircle
-  }
-}
 
 /**
  * Alert Component.
@@ -86,29 +24,27 @@ const variants = {
  * @param {object} props
  */
 export const Alert = (props) => {
+
+  // Define variants.
+  const variants = {
+    error: <AlertError {...props} />,
+    info: <AlertInfo {...props} />,
+    success: <AlertSuccess {...props} />,
+    warning: <AlertWarning {...props} />
+  }
+
+  // Switch to variant.
+  if (props.variant && variants[props.variant] !== 'undefined') {
+    return variants[props.variant]
+  }
+
   // Removes the alert from display when clicked.
   const [isDismissed, setDismissed] = useState(false)
 
-  // Merge with passed in props.
-  let classes = Object.assign(defaultClasses, props.classes)
-  // Merge with variant setting from props.
-  classes = Object.assign(classes, variants[props.variant])
-
-  // The dismiss button component.
-  // Toggle to dark based on variant.
-  const dismissDarkVariant = ['warning']
-  let dismissVariant = 'light'
-  if (
-    props.variant === undefined ||
-    dismissDarkVariant.includes(props.variant)
-  ) {
-    dismissVariant = 'dark'
-  }
-
   // DismissButtton component.
   const dismiss = (
-    <div className={classes.dismissButtonWrapper}>
-      <DismissButton variant={dismissVariant} callback={setDismissed} />
+    <div className='su-order-3 su-rs-ml-1 su-h-full su-items-end su-flex-shrink su-text-right su-w-full sm:su-w-auto'>
+      <DismissButton variant='dark' callback={setDismissed} />
     </div>
   )
 
@@ -119,24 +55,24 @@ export const Alert = (props) => {
 
   // The goods!
   return (
-    <div className={classes.wrapper}>
+    <div className='su-alert'>
       <div className='su-cc su-flex su-flex-wrap sm:su-items-center'>
         {props.dismiss && dismiss}
-        <div className={classes.headerWrapper}>
-          <span className={classes.headerIcon}>
+        <div className='su-order-1 su-rs-mr-1 su-flex-shrink su-mb-4 xs:su-w-full lg:su-w-max'>
+          <span className=''>
             {props.icon ?? (
               <FontAwesomeIcon
-                icon={classes.icon}
-                className={classes.iconClass}
+                icon={faBell}
+                className='su-mr-2 su-inline-block su-max-w-xs'
               />
             )}
           </span>
-          <span className={classes.label}>{props.label ?? 'Information'}</span>
+          <span className={'su-inline-block su-uppercase su-font-semibold su-text-170rem su-h-full ' + styles.label}>{props.label ?? 'Information'}</span>
         </div>
-        <div className={classes.bodyWrapper}>
-          <h3 className={classes.bodyHeading}>{props.heading}</h3>
-          <div className={classes.body}>{props.children}</div>
-          <div className={classes.footerWrapper}>{props.footer}</div>
+        <div className={styles.alertBodyWrapperDark}>
+          <h3>{props.heading}</h3>
+          <div>{props.children}</div>
+          <div className='su-rs-mt-0'>{props.footer}</div>
         </div>
       </div>
     </div>
@@ -148,14 +84,12 @@ export const Alert = (props) => {
  */
 
 Alert.propTypes = {
-  // An array of classes to merge with the classes array.
-  classes: PropTypes.object,
   // The primary content
   children: PropTypes.node,
   // FA Icon name.
   icon: PropTypes.element,
   // One of the default variant sets.
-  variant: PropTypes.oneOf(Object.keys(variants)),
+  variant: PropTypes.oneOf(['error', 'info', 'success', 'warning']),
   // Show the dismiss button or not.
   dismiss: PropTypes.bool,
   // Add a label.
